most_frequent_google_searches/requirements.txt
numpy 
pandas 
matplotlib
wordcloud
Pillow
